requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==21.6b0
bottle==0.12.19
cffi==1.14.5
click==8.0.1
dnspython==1.16.0
dotsi==0.0.3
hupper==1.10.3
importlib-metadata==4.6.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
pymongo==3.11.4
python-dotenv==0.18.0
regex==2021.4.4
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
vf==0.0.2
waitress==2.0.0
zipp==3.4.1
